[ADD] GCN

+++ train.py
@@ -1,6 +1,7 @@
 # TODO: - Create multiple model for GNN
 import argparse
 
+import torch
 from dgl import add_self_loop
 from torch.nn.functional import cross_entropy, relu
 from torch.optim import Adam
@@ -32,7 +33,8 @@ def train(
     best_val_f1 = 0
     best_test_f1 = 0
 
-    features = g.ndata["features"]
+    # FIXME: Convert this in graph creattion to float32
+    features = g.ndata["features"].to(torch.float32)
     labels = g.ndata["label"]
     # label_binarizer = LabelBinarizer()
     # label_binarizer.fit(range(max(labels) + 1))
@@ -146,7 +148,7 @@ if __name__ == "__main__":
         relu,
     ).to("cuda")
     # TODO: here sometime float some time double
-    model.double()
+    model.float()
     edge_weight = graph_train.edata["weight"].double().to("cuda")
 
     train_list, val_list, test_list, loss, loss_val, loss_test = train(
